pyproject.toml
[tool.poetry]
package-mode = false
name = "rob530"
version = "0.1.0"
description = ""
authors = ["marshallvielmetti <mvielmetti@outlook.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2.1"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
